Add: article-user, article-comment 관계 실습파일

+++ Django/kdt_Django/10_relationsships/relationships_practice/articles/views.py
@@ -22,15 +22,17 @@ def detail(request,pk):
         'comment_form' : comment_form,
         'comments' : comments,
     }
-
     return render(request, 'articles/detail.html', context)
+
 
 @login_required
 def create(request):
     if request.method == 'POST':
         form = ArticleForm(request.POST)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -42,9 +44,11 @@ def create(request):
 
 @login_required
 def delete(request, pk):
+    # 삭제를 요청하는 자 vs 게시글의 작성자를 비교
     article = Article.objects.get(pk=pk)
-    article.delete()
-
+    if request.user == article.user:
+        article.delete()
+    
     return redirect('articles:index')
 
 
@@ -52,15 +56,17 @@ def delete(request, pk):
 @login_required
 def update(request, pk):
     article = Article.objects.get(pk=pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-
-        if form.is_valid():
-            form.save()
-            return redirect('articles:index')
+    # 수정을 요청하는 자 vs 게시글의 작성자를 비교
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:index')
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
-        
+        return redirect('articles:index')
     context = {
         'article' : article,
         'form' : form
@@ -68,6 +74,7 @@ def update(request, pk):
     return render(request, 'articles/edit.html', context)     
 
 
+@login_required
 def comment_create(request, pk):
     # 몇 번 게시글인지 조회
     article = Article.objects.get(pk=pk)
@@ -75,9 +82,10 @@ def comment_create(request, pk):
     comment_form = CommentForm(request.POST)
     # 유효성 검증
     if comment_form.is_valid():
-        # commit을 flase로 주면 인스턴스는 반환하면서도 DB에 레코드는 작성하지 않도록 함
+        # commit을 False로 주면 인스턴스는 반환하면서도 DB에 레코드는 작성하지 않도록 함
         comment = comment_form.save(commit=False)
         comment.article = article
+        comment.user = request.user
         comment.save()
         return redirect('articles:detail', pk)
     context = {
@@ -89,9 +97,12 @@ def comment_create(request, pk):
     return render(request, 'articles/detail.html', context)
 
 
+@login_required
 def comment_delete(request, article_pk, comment_pk):
     # 삭제할 댓글을 조회
     comment = Comment.objects.get(pk = comment_pk)
-    # 댓글 삭제
-    comment.delete()
+    # 삭제를 요청하는 자 vs 댓글의 작성자를 비교
+    if comment.user == request.user:
+        # 댓글 삭제
+        comment.delete()
     return redirect('articles:detail', article_pk)
